move example fixtures under tests
