+++ tests/examples/pyminifier.py
