+++ tests/examples/pyminifier.pymini.py
