+++ tests/examples/pyminify.py
