+++ tests/examples/pyminify.pymini.py
